rename addl skillfunctionaltests refs to components
