+++ tests/functional/Tests/ComponentsFunctionalTests/Common/TestCaseBuilder.cs
@@ -5,7 +5,7 @@ using System;
 using System.Collections.Generic;
 using TranscriptTestRunner;
 
-namespace SkillFunctionalTests.Common
+namespace ComponentsFunctionalTests.Common
 {
     public class TestCaseBuilder
     {
