+++ tests/functional/Tests/ComponentsFunctionalTests/EmptyBot/TestScripts/EmptyBot.json
