+++ tests/functional/Tests/ComponentsFunctionalTests/EmptyBot/TestScripts/EmptyBot.transcript
